backend/requirements.txt
Django==4.2.11
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
psycopg2-binary==2.9.11
python-dotenv==1.0.1
Pillow==10.4.0
WeasyPrint==62.3
django-filter==23.5
